webapp/requirements.txt
shinylive
astral==2.2
